added task

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -16,7 +16,7 @@
 
 int get_bit(unsigned long int n, unsigned int k)
 {
-	unsigned int i = 0, g = n;
+	unsigned long int i = 0, g = n;
 
 	if (n == 0)
 	{
